requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.5.0
httpx==0.27.0
apscheduler==3.10.4
boto3==1.34.0
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.6
ruff==0.4.4
slowapi==0.1.9
